+++ demisto_sdk/commands/content_graph/content_graph_commands.py
@@ -7,6 +7,7 @@ import demisto_sdk.commands.content_graph.neo4j_service as neo4j_service
 from demisto_sdk.commands.common.constants import MarketplaceVersions
 from demisto_sdk.commands.common.git_util import GitUtil
 from demisto_sdk.commands.common.tools import download_content_graph
+from demisto_sdk.commands.content_graph.common import NEO4J_DATABASE_HTTP
 from demisto_sdk.commands.content_graph.content_graph_builder import ContentGraphBuilder
 from demisto_sdk.commands.content_graph.interface.graph import ContentGraphInterface
 
@@ -33,6 +34,9 @@ def create_content_graph(
     if output_path:
         output_path = output_path / marketplace.value
     content_graph_interface.export_graph(output_path)
+    logger.info(
+        f"Successfully created the content graph. UI representation is available at {NEO4J_DATABASE_HTTP}"
+    )
 
 
 def update_content_graph(
@@ -59,23 +63,28 @@ def update_content_graph(
     if packs_to_update is None:
         packs_to_update = []
     builder = ContentGraphBuilder(content_graph_interface)
-    if not imported_path and not use_current:
-        # getting the graph from remote, so we need to clean the import dir
+    if not use_current:
         content_graph_interface.clean_import_dir()
-        extract_remote_import_files(content_graph_interface, builder)
+        if not imported_path:
+            # getting the graph from remote, so we need to clean the import dir
+            extract_remote_import_files(content_graph_interface, builder)
 
     if use_git and (commit := content_graph_interface.commit):
         packs_to_update.extend(GitUtil().get_all_changed_pack_ids(commit))
 
     content_graph_interface.import_graph(imported_path)
 
-    logger.info(f"Updating the following packs: {packs_to_update}")
+    packs_str = "\n".join([f"- {p}" for p in packs_to_update])
+    logger.info(f"Updating the following packs:\n{packs_str}")
     builder.update_graph(packs_to_update)
     if dependencies:
         content_graph_interface.create_pack_dependencies()
     if output_path:
         output_path = output_path / marketplace.value
     content_graph_interface.export_graph(output_path)
+    logger.info(
+        f"Successfully updated the content graph. UI representation is available at {NEO4J_DATABASE_HTTP}"
+    )
 
 
 def extract_remote_import_files(